gpu_instance/pyproject.toml
[project]
name = "gpu-instance"
version = "0.1.0"
description = "GPU transcription worker using faster-whisper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.24",
    "botocore[crt]>=1.42.24",
    "faster-whisper>=1.2.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
transcribe = "gpu_instance.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gpu_instance"]
